requirements.txt
# Core PDF processing
pdfplumber>=0.11.0
pymupdf>=1.24.0
Pillow>=10.2.0

# Schema validation
pydantic>=2.0.0

# Web framework
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Vision LLM backends
ollama>=0.2.0
openai>=1.0.0          # Used for DeepSeek API (OpenAI-compatible)

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
httpx>=0.27.0          # For FastAPI test client
